update user controller

+++ user/user.service.js
@@ -6,8 +6,6 @@ const User = db.User;
 
 module.exports = {
     authenticate,
-    getAll,
-    getById,
     create,
     update,
     delete: _delete
@@ -15,7 +13,7 @@ module.exports = {
 
 async function authenticate({email, password}) {
     const user = await User.findOne({email});
-    if (user && bcrypt.compareSync(password, user.hash)) {
+    if (user && bcrypt.compareSync(password, user.password)) {
         const token = jwt.sign({sub: user.id}, config.secret, {expiresIn: '7d'});
         return {
             ...user.toJSON(),
@@ -24,44 +22,36 @@ async function authenticate({email, password}) {
     }
 }
 
-async function getAll() {
-    return await User.find();
-}
-
-async function getById(id) {
-    return await User.findById(id);
-}
-
 async function create(userParam) {
     console.log(userParam);
     // validate
     if (await User.findOne({email: userParam.email})) {
-        throw 'Email "' + userParam.username + '" is already taken';
+        throw 'Email "' + userParam.name + '" is already taken';
     }
 
     const user = new User(userParam);
 
     // hash password
     if (userParam.password) {
-        user.hash = bcrypt.hashSync(userParam.password, 10);
+        user.password = bcrypt.hashSync(userParam.password, 10);
     }
 
     // save user
     await user.save();
 }
 
-async function update(id, userParam) {
-    const user = await User.findById(id);
+async function update(email, userParam) {
+    const user = await User.findOne({email: email});
 
     // validate
-    if (!user) throw 'User not found';
-    if (user.username !== userParam.username && await User.findOne({username: userParam.username})) {
-        throw 'Username "' + userParam.username + '" is already taken';
+    if (!user) throw 'User '+email+ ' not found';
+    if (user.email !== userParam.email && await User.findOne({email: userParam.email})) {
+        throw 'Email "' + userParam.email + '" is already taken';
     }
 
     // hash password if it was entered
     if (userParam.password) {
-        userParam.hash = bcrypt.hashSync(userParam.password, 10);
+        userParam.password = bcrypt.hashSync(userParam.password, 10);
     }
 
     // copy userParam properties to user
@@ -71,5 +61,9 @@ async function update(id, userParam) {
 }
 
 async function _delete(email) {
+    const user = await User.findOne({email: email});
+
+    // validate
+    if (!user) throw 'User '+email+ ' not found';
     await User.deleteOne({email: email});
 }
